+++ frontend/src/components/google-maps/marker/tools/index.ts
